Better testing, removal of negative std
The ANFIS model used to have negative stds, but that has been changed. Ideally in the future, the means and the stds can be a parameter to be changed, that way the distribution for them can be updated.

Also, I would like to be able to add more than just Gaussian membership functions. For example, adding in triangular, sigmoid, and maybe trapezoidal.

Other things to add to the models would be Dueling, which shouldn't be too hard