Better ANFIS
Okay there was a big problem with the defuzzification of the network. Now it has been tested on learning a quadratic function and performed fairly well.

Because of the structure change, had to update the parameters to the networks to match parameter count.

Also, changed how the input/output dims are passed to each network to make it more generalize-able.