ANFIS and DQN comparison
Seems that depending on the seed, the ANFIS may learn faster than the DQN, and vice versa. It seems that with hyperparameter tuning it could be a good contender, especially for some more complicated environments. Would be good to test more, but this is about the amount of work necessary for the paper.