Triangular membership functions and plotting the rules
There are some improvements that can be made to triangular membership functions (and Gaussian as well to save some memory space). This will be updated and applied at a future point.

Plotting is currently very rudimentary, and seems okay. It also could probably be improved upon with the same changes made to the membership functions. Also, having the ability to swap between different layers might be extremely useful (perhaps a 3d plot of some kind? Or using a different plotting service altogether).